requirements.txt
asgiref==3.3.4
Django==3.1.7
django-appconf==1.0.4
django-bootstrap==0.2.4
django-bootstrap-v5==1.0.0
django-cors-headers==3.10.0
django-extensions==3.1.3
django-imagekit==4.0.2
django-seed==0.3.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
pytz==2021.1
sqlparse==0.4.1
backcall==0.2.0
beautifulsoup4==4.9.3
certifi==2021.10.8
decorator==4.4.2
